+++ src/main/java/com/moilioncircle/redis/replicator/cmd/impl/FlushAllCommand.java
@@ -25,12 +25,23 @@ public class FlushAllCommand extends AbstractCommand {
     private static final long serialVersionUID = 1L;
     
     private boolean async;
+    private boolean sync;
 
     public FlushAllCommand() {
     }
     
-    public FlushAllCommand(final boolean async) {
+    public FlushAllCommand(boolean async) {
+        this(async, false);
+    }
+    
+    /**
+     * @since 3.5.2
+     * @param async async
+     * @param sync sync
+     */
+    public FlushAllCommand(boolean async, boolean sync) {
         this.async = async;
+        this.sync = sync;
     }
 
     public boolean isAsync() {
@@ -39,5 +50,21 @@ public class FlushAllCommand extends AbstractCommand {
 
     public void setAsync(boolean async) {
         this.async = async;
+    }
+    
+    /**
+     * @since 3.5.2
+     * @return sync
+     */
+    public boolean isSync() {
+        return sync;
+    }
+    
+    /**
+     * @since 3.5.2
+     * @param sync sync
+     */
+    public void setSync(boolean sync) {
+        this.sync = sync;
     }
 }