+++ src/main/java/com/moilioncircle/redis/replicator/cmd/impl/MinId.java
@@ -1,5 +1,5 @@
 /*
- * Copyright 2016-2018 Leon Chen
+ * Copyright 2016-2017 Leon Chen
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -16,46 +16,42 @@
 
 package com.moilioncircle.redis.replicator.cmd.impl;
 
+import java.io.Serializable;
+
 /**
  * @author Leon Chen
- * @since 2.1.0
+ * @since 3.5.2
  */
-public class LPopCommand extends GenericKeyCommand {
-
+public class MinId implements Serializable {
+    
     private static final long serialVersionUID = 1L;
     
-    private Count count;
-
-    public LPopCommand() {
-    }
-
-    public LPopCommand(byte[] key) {
-        this(key, null);
+    private boolean approximation;
+    
+    private byte[] id;
+    
+    public MinId() {
+        
     }
     
-    /**
-     * @since 3.5.2
-     * @param key key
-     * @param count count
-     */
-    public LPopCommand(byte[] key, Count count) {
-        super(key);
-        this.count = count;
+    public MinId(boolean approximation, byte[] id) {
+        this.approximation = approximation;
+        this.id = id;
     }
     
-    /**
-     * @since 3.5.2
-     * @return count
-     */
-    public Count getCount() {
-        return count;
+    public boolean isApproximation() {
+        return approximation;
     }
     
-    /**
-     * @since 3.5.2
-     * @param count count
-     */
-    public void setCount(Count count) {
-        this.count = count;
+    public void setApproximation(boolean approximation) {
+        this.approximation = approximation;
+    }
+    
+    public byte[] getId() {
+        return id;
+    }
+    
+    public void setId(byte[] id) {
+        this.id = id;
     }
 }