+++ src/main/java/com/moilioncircle/redis/replicator/cmd/parser/FlushAllParser.java
@@ -30,10 +30,13 @@ public class FlushAllParser implements CommandParser<FlushAllCommand> {
     @Override
     public FlushAllCommand parse(Object[] command) {
         boolean isAsync = false;
+        boolean isSync = false;
         if (command.length == 2 && isEquals(toRune(command[1]), "ASYNC")) {
             isAsync = true;
+        } else if (command.length == 2 && isEquals(toRune(command[1]), "SYNC")) {
+            isSync = true;
         }
-        return new FlushAllCommand(isAsync);
+        return new FlushAllCommand(isAsync, isSync);
     }
 
 }