+++ src/main/java/com/moilioncircle/redis/replicator/cmd/parser/RPopParser.java
@@ -17,9 +17,11 @@
 package com.moilioncircle.redis.replicator.cmd.parser;
 
 import com.moilioncircle.redis.replicator.cmd.CommandParser;
+import com.moilioncircle.redis.replicator.cmd.impl.Count;
 import com.moilioncircle.redis.replicator.cmd.impl.RPopCommand;
 
 import static com.moilioncircle.redis.replicator.cmd.CommandParsers.toBytes;
+import static com.moilioncircle.redis.replicator.cmd.CommandParsers.toInt;
 
 /**
  * @author Leon Chen
@@ -29,7 +31,11 @@ public class RPopParser implements CommandParser<RPopCommand> {
     @Override
     public RPopCommand parse(Object[] command) {
         byte[] key = toBytes(command[1]);
-        return new RPopCommand(key);
+        Count count = null;
+        if (command.length == 3) {
+            count = new Count(toInt(command[2]));
+        }
+        return new RPopCommand(key, count);
     }
 
 }